Remove Bitcoin
Bitcoin has turned out to be little more than a digital asset for investment and speculation.

Does it work as peer-to-peer electronic cash? Yeah, if we define "works" as "technically, you can issue and receive transactions", but the UX is awful and the risks unreasonably high - not to mention that the network does not scale beyond a dozen or so TXns per second.

Bitcoin is narrowly a success at what advocates say it can do, but that success will likely never translate into "everyday people who don't care about the engineering and technical aspects will use this for anything". Bitcoin is a success and failure.